Refactor browser cards presenter into modular apps
